# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cadastro de Endereços",
        "description": "API realizada durante o bootcampo de PHP com Laravel da Squadra",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/bairro": {
            "get": {
                "tags": [
                    "Bairros"
                ],
                "summary": "Retorna uma lista de bairros cadastrados",
                "description": "Retorna um Array com todos os bairros cadastrados",
                "operationId": "8c4adb1c5a368105c1534e78e5779711",
                "responses": {
                    "200": {
                        "description": "Array com bairros"
                    },
                    "503": {
                        "description": "JSON com mensagem de que não foi possível pesquisar um determinado bairro."
                    }
                }
            },
            "post": {
                "tags": [
                    "Bairros"
                ],
                "summary": "Retorna o bairro cadastrado",
                "description": "Armazena um novo bairro.",
                "operationId": "536fd30f2762a5ebd62dfb94e6540145",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo_municipio": {
                                        "type": "integer"
                                    },
                                    "nome": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codigo_municipio": 5,
                                    "nome": "CENTRO",
                                    "status": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bairro cadastrado com sucesso."
                    },
                    "503": {
                        "description": "Não foi possível cadastrar o bairro."
                    }
                }
            }
        },
        "/bairro/{id}": {
            "get": {
                "tags": [
                    "Bairros"
                ],
                "summary": "Retorna um bairro cadastrado",
                "description": "Retorna um JSON com o bairro pesquisado.",
                "operationId": "e560ec2f8471072f1388e0295163271b",
                "responses": {
                    "200": {
                        "description": "JSON com o bairro"
                    },
                    "503": {
                        "description": "JSON com mensagem de que não foi possível encontrar um determinado bairro."
                    }
                }
            },
            "put": {
                "tags": [
                    "Bairros"
                ],
                "summary": "Retorna o bairro atualizado.",
                "description": "Atualiza um bairro",
                "operationId": "4d2a5a09311f9b7094d3db9b90e6a9b5",
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Dados da requisição",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do bairro que será atualizado.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bairro atualizado com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar o bairro."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bairros"
                ],
                "summary": "Retorna um array vazio.",
                "description": "Apaga um bairro",
                "operationId": "f2b6ad517564df22c631f44aad83429e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do bairro que será apagado.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bairro apagado com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar o bairro."
                    }
                }
            }
        },
        "/municipio": {
            "get": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Retorna uma lista de municípios cadastrados",
                "description": "Retorna um Array com todos os municípios cadastrados",
                "operationId": "adc28966a21c645f4ce8e50eabfd8d1a",
                "responses": {
                    "200": {
                        "description": "Array com municípios"
                    },
                    "503": {
                        "description": "JSON com mensagem de que não foi possível pesquisar um determinado município."
                    }
                }
            },
            "post": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Retorna o município cadastrado",
                "description": "Armazena um novo Município.",
                "operationId": "91b057fad8e8461b12343a98512f2b9a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo_uf": {
                                        "type": "integer"
                                    },
                                    "nome": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codigo_uf": "1",
                                    "nome": "TAGUATINGA",
                                    "status": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Município cadastrado com sucesso."
                    },
                    "503": {
                        "description": "Não foi possível cadastrar o Município."
                    }
                }
            }
        },
        "/municipio/{id}": {
            "get": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Retorna um município cadastrado",
                "description": "Retorna um JSON com todos o município pesquisado.",
                "operationId": "6dc36a230e96258a4157e6453d6dc0ea",
                "responses": {
                    "200": {
                        "description": "JSON com o município"
                    },
                    "503": {
                        "description": "JSON com mensagem de que não foi possível encontrar um determinado município."
                    }
                }
            },
            "put": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Retorna um município atualizado.",
                "description": "Atualiza um Município",
                "operationId": "df4fc756d73faa0fee7b58547b44093e",
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Dados da requisição",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do município que será atualizado.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Município atualizado com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar o município."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Retorna um array vazio.",
                "description": "Apaga uma município",
                "operationId": "e03769af96b9def5e3117e0288b6e610",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do município que será apagado.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Município apagado com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar o município."
                    }
                }
            }
        },
        "/uf": {
            "get": {
                "tags": [
                    "UFs"
                ],
                "summary": "Retorna uma lista de UFs",
                "description": "Retorna um Array com todos os UFs cadastrados",
                "operationId": "ea813f44177e7db9b4217a80c408fc24",
                "responses": {
                    "200": {
                        "description": "Array com UFs"
                    },
                    "503": {
                        "description": "JSON com mensagem de que não foi possível pesquisar uma determinada UF"
                    }
                }
            },
            "post": {
                "tags": [
                    "UFs"
                ],
                "summary": "Retorna a UF cadastrada",
                "description": "Armazena uma nova UF",
                "operationId": "e7dfd1453a879b3075d683780f40c7ea",
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Dados da requisição",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UF cadastrada com sucesso."
                    },
                    "503": {
                        "description": "Não foi possível cadastrar a UF."
                    }
                }
            }
        },
        "/uf/{id}": {
            "get": {
                "tags": [
                    "UFs"
                ],
                "summary": "Retorna uma lista de UFs",
                "description": "Retorna um Array com todos os UFs cadastrados",
                "operationId": "33797d11424777ee2ee7ea599d725812",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da UF",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna UF em formato JSON"
                    },
                    "503": {
                        "description": "Json com mensagem de que não foi possível encontrar uma determinada UF."
                    }
                }
            },
            "put": {
                "tags": [
                    "UFs"
                ],
                "summary": "Retorna a UF atualizada",
                "description": "Atualiza uma nova UF",
                "operationId": "c84d08baa602d585f625d3722d609e68",
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "Dados da requisição",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da UF que será atualizada.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UF atualizada com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar a UF."
                    }
                }
            },
            "delete": {
                "tags": [
                    "UFs"
                ],
                "summary": "Retorna um array vazio.",
                "description": "Apaga uma nova UF",
                "operationId": "21c6c11e0fd5e3e53c44f8541496116a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da UF que será apagada.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UF apagada com sucesso."
                    },
                    "404": {
                        "description": "Não foi possível encontrar a UF."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Cadastro de Endereços",
            "description": "API Endpoints da API de Cadastro de Endereços"
        }
    ]
}